+++ controllers/report/getCategoryPerMonthStatistics.js
@@ -1,17 +1,25 @@
 const asyncHandler = require('express-async-handler')
-
-const shortPerMonthStatisticsServises = require('../../services/shortPerMonthStatistics')
+const reportServises = require('../../services/report')
 
 const controller = asyncHandler(async (req, res) => {
-  const { date } = req.query
+  const { date, transactionType } = req.query
   if (!date) {
     res.status(400)
     throw new Error(`Date ${date} is not supported`)
   }
+
+  const testType = ['income', 'expenses'].find(item => item === transactionType)
+
+  if (!testType) {
+    res.status(400)
+    throw new Error(`Transaction type ${transactionType} is not supported`)
+  }
   const currentDate = new Date(date)
   const month = currentDate.getMonth().toString()
   const year = currentDate.getFullYear().toString()
-  const result = await shortPerMonthStatisticsServises(month, year)
+  const type = transactionType.toString()
+
+  const result = await reportServises.categoryPerMonthStatistics(month, year, type)
 
   if (!result) {
     res.status(404)
