+++ controllers/report/index.js
@@ -1,0 +1,11 @@
+const getShortStatistics = require('./getShortStatistics')
+const getPerMonthStatistics = require('./getPerMonthStatistics')
+const getCategoryPerMonthStatistics = require('./getCategoryPerMonthStatistics')
+
+const report = {
+  getShortStatistics,
+  getPerMonthStatistics,
+  getCategoryPerMonthStatistics
+}
+
+module.exports = report
