+++ services/report/index.js
@@ -1,0 +1,11 @@
+const shortStatistics = require('./shortStatistics')
+const shortPerMonthStatistics = require('./shortPerMonthStatistics')
+const categoryPerMonthStatistics = require('./categoryPerMonthStatistics')
+
+const reportServises = {
+  shortStatistics,
+  shortPerMonthStatistics,
+  categoryPerMonthStatistics
+}
+
+module.exports = reportServises
