+++ services/report/shortPerMonthStatistics.js
