+++ services/report/shortStatistics.js
